# openapi.yaml
openapi: 3.0.0
info:
  title: 'Cube-3 API'
  contact:
    email: leo.paillard@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://cyber-dodo.fr:8080/api/v1'
    description: 'Cube-3 API'
paths:
...
  /task:
    post:
      tags:
        - Task
      summary: 'Add a new task'
      description: 'Add a new task to the database'
      operationId: 7f1ce22b98f5151d7ca2602801db555e
      requestBody:
        description: 'Task object that needs to be added to the database'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: 'Task created'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'User, Project or TaskStatus not found'
        '409':
          description: 'Task already exists'
        '500':
          description: 'Internal server error'
  /task/all:
    get:
      tags:
        - Task
      summary: 'Get all tasks'
      description: 'Returns all tasks'
      operationId: d85a7cb58dd5f645ce501b6963fb9a32
      responses:
        '200':
          description: 'Tasks found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '500':
          description: 'Internal server error'
  '/task/user/{id}':
    get:
      tags:
        - Task
      summary: 'Get all tasks by user'
      description: 'Returns all tasks assigned to a user'
      operationId: 26b84f9fd60a949c9d2e60cb3a178f24
      parameters:
        -
          name: id
          in: path
          description: 'User ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Tasks found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '404':
          description: 'User not found'
        '500':
          description: 'Internal server error'
  '/task/project/{id}':
    get:
      tags:
        - Task
      summary: 'Get all tasks by project'
      description: 'Returns all tasks assigned to a project'
      operationId: 41dbadc7c0530c7542fd1f8a641466fe
      parameters:
        -
          name: id
          in: path
          description: 'Project ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Tasks found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '500':
          description: 'Internal server error'
  '/task/{id}':
    get:
      tags:
        - Task
      summary: 'Get task by id'
      description: 'Returns a task by id'
      operationId: 82adb562defb983aa7f09a7b1ea8224a
      parameters:
        -
          name: id
          in: path
          description: 'Task ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Task found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: 'Task not found'
        '500':
          description: 'Internal server error'
    put:
      tags:
        - Task
      summary: 'Update a task by ID'
      description: 'Update a task by ID'
      operationId: bd75423333b743d05991e8a34133a7d5
      parameters:
        -
          name: id
          in: path
          description: 'ID of the task to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Task object that needs to be updated'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: 'Task updated'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'Task not found'
        '409':
          description: 'Task already exists'
        '500':
          description: 'Internal server error'
    delete:
      tags:
        - Task
      summary: 'Deletes a task by ID'
      description: 'Deletes a task by ID'
      operationId: ca22588758e816deca16255f28fd4bef
      parameters:
        -
          name: id
          in: path
          description: 'ID of the task to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Task deleted'
        '404':
          description: 'Task not found'
        '500':
          description: 'Internal server error'
...
components:
  schemas:
...
    CompanyResponse:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Cube 3'
        address:
          type: string
          example: '1, rue de la paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        slogan:
          type: string
          example: 'The best company ever'
        logoPath:
          type: string
          example: cube3.png
        license:
          $ref: '#/components/schemas/LicenseResponse'
        language:
          type: string
          example: fr
        licenseExpirationDate:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    CustomerResponse:
      properties:
        id:
          type: integer
          example: '1'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john-doe@gmail.com
        address:
          type: string
          example: '1, rue de la Paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        company:
          $ref: '#/components/schemas/CompanyResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
        status:
          $ref: '#/components/schemas/CustomerStatusResponse'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    CustomerStatusResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'CustomerStatus 1'
        description:
          type: string
          example: 'This is the first customerStatus'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    LicenseResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: basic
        description:
          type: string
          example: 'basic license for 1 user'
        price:
          type: integer
          example: 100
        maxUsers:
          type: integer
          example: 1
        validityPeriod:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
      type: object
...
    ProjectResponse:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Project 1'
        description:
          type: string
          example: 'This is the first project'
        company:
          $ref: '#/components/schemas/CompanyResponse'
        creator:
          $ref: '#/components/schemas/UserResponse'
        customer:
          $ref: '#/components/schemas/CustomerResponse'
        projectStatus:
          type: integer
          example: '1'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    RoleResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Admin
        description:
          type: string
          example: Administrator
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    TaskRequest:
      required:
        - title
        - description
        - location
        - dueDate
        - project
        - user
        - taskStatus
      properties:
        title:
          type: string
          example: 'Task 1'
        description:
          type: string
          example: 'This is the first task'
        location:
          type: string
          example: 'Location 1'
        dueDate:
          type: string
          example: '2021-01-01'
        project:
          type: integer
          example: 1
        user:
          type: integer
          example: 1
        taskStatus:
          type: integer
          example: 1
        taskType:
          type: integer
          example: 1
      type: object
    TaskResponse:
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Task 1'
        description:
          type: string
          example: 'This is the first task'
        location:
          type: string
          example: 'Location 1'
        dueDate:
          type: string
          example: '2021-01-01'
        project:
          $ref: '#/components/schemas/ProjectResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
        taskStatus:
          $ref: '#/components/schemas/TaskStatusResponse'
        taskType:
          $ref: '#/components/schemas/TaskTypeResponse'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    TaskStatusResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'TaskStatus 1'
        description:
          type: string
          example: 'This is the first taskStatus'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    TaskTypeResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'TaskStatus 1'
        description:
          type: string
          example: 'This is the first taskStatus'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    UserResponse:
      properties:
        id:
          type: integer
          example: '1'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@gmail.com
        job:
          type: string
          example: Developer
        phone:
          type: string
          example: '0123456789'
        role:
          $ref: '#/components/schemas/RoleResponse'
        company:
          type: integer
          example: '1'
        createdAt:
          type: string
          format: date-time
          example: '2021-03-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-03-01 00:00:00'
      type: object
...
  securitySchemes:
    bearer:
      type: http
      bearerFormat: JWT